ascom_scripts/decrypt_requirements.txt
aes==1.0.0
certifi==2020.6.20
chardet==3.0.4
idna==2.10
install==1.3.3
Naked==0.1.31
pkcs7==0.1.2
pycryptodome==3.9.8
pycryptodomex==3.9.8
PyYAML==5.3.1
requests==2.24.0
shellescape==3.8.1
urllib3==1.25.10
